+++ tests/e2e/manifest-lint/baseline.txt
@@ -79,7 +79,6 @@ create	tests/e2e/single/negative_test.go	RunInstances	#2
 create	tests/e2e/single/negative_test.go	RunInstances	#3
 create	tests/e2e/single/negative_test.go	RunInstances	#4
 create	tests/e2e/single/snapshot_test.go	CreateSnapshot	#1
-create	tests/e2e/single/volume_durability_test.go	CreateVolume	#1
 create	tests/e2e/single/volume_test.go	CreateVolume	#1
 create	tests/e2e/single/vpc_test.go	CreateInternetGateway	#1
 create	tests/e2e/single/vpc_test.go	CreateRouteTable	#1
