+++ tests/e2e/single/instance_test.go
@@ -4,10 +4,8 @@ package single
 
 import (
 	"bytes"
-	"net"
 	"os/exec"
 	"strconv"
-	"strings"
 	"testing"
 
 	"github.com/aws/aws-sdk-go/aws"
@@ -65,68 +63,6 @@ func runInstanceMetadata(t *testing.T, fix *Fixture) {
 		"image", aws.StringValue(inst.ImageId),
 		"bdms", len(inst.BlockDeviceMappings),
 	)
-}
-
-// runSSHProbe waits for SSH to become reachable, then runs `id`,
-// `lsblk`, and `hostname` against the guest to confirm the VM booted and
-// presents the expected root volume. Maps to run-e2e.sh ~381–463.
-func runSSHProbe(t *testing.T, fix *Fixture) {
-	harness.Phase(t, "Single — SSH Connectivity Phase 5a-ii — SSH Connectivity & Volume Verification Volume Verification")
-
-	inst, rootVolumeID := needInstance(t, fix)
-	_, keyPath := needKeyPair(t, fix)
-	instanceID := aws.StringValue(inst.InstanceId)
-
-	host, port := harness.InstancePublicSSHHost(t, inst)
-	harness.Detail(t, "ssh_host", host, "ssh_port", port)
-
-	addr := net.JoinHostPort(host, strconv.Itoa(port))
-	harness.Step(t, "waiting for SSH handshake %s", addr)
-	waitForSSHHandshake(t, host, port, keyPath)
-	_ = addr
-
-	tgt := harness.SSHTarget{User: "ubuntu", Host: host, Port: port, KeyPath: keyPath}
-
-	harness.Step(t, "ssh id")
-	idOut := runSSH(t, tgt, "id")
-	assert.Containsf(t, idOut, "ubuntu", "ssh id should report ubuntu\n%s", idOut)
-
-	harness.Step(t, "lsblk root-volume cross-check vs API")
-	guestGiB := harness.LsblkRootGiB(t, tgt)
-
-	vols, err := fix.AWS.EC2.DescribeVolumes(&ec2.DescribeVolumesInput{
-		VolumeIds: []*string{aws.String(rootVolumeID)},
-	})
-	require.NoError(t, err, "describe-volumes %s", rootVolumeID)
-	require.NotEmpty(t, vols.Volumes, "no volume for %s", rootVolumeID)
-	apiGiB := int(aws.Int64Value(vols.Volumes[0].Size))
-
-	// lsblk rounds down — bash treats equality as required, but on backing
-	// stores where the VM's view is a hair under the API size the rounding
-	// loses 1 GiB. Allow ±1 to match the bash intent without flaking on it.
-	diff := guestGiB - apiGiB
-	if diff < 0 {
-		diff = -diff
-	}
-	assert.LessOrEqualf(t, diff, 1,
-		"root volume size mismatch: guest=%d GiB api=%d GiB", guestGiB, apiGiB)
-	harness.Detail(t, "guest_gib", guestGiB, "api_gib", apiGiB)
-
-	harness.Step(t, "ssh hostname")
-	hn := strings.TrimSpace(runSSH(t, tgt, "hostname"))
-	// Bash uses `spinifex-vm-<first 8 hex chars of instance ID>` and treats
-	// a missing prefix as a non-fatal warning. Replicate the soft check via
-	// t.Logf rather than asserting — the spx hostname format is not part of
-	// the EC2 surface contract.
-	shortID := strings.TrimPrefix(instanceID, "i-")
-	if len(shortID) > 8 {
-		shortID = shortID[:8]
-	}
-	if strings.Contains(hn, shortID) {
-		harness.Detail(t, "hostname", hn, "matches_short_id", shortID)
-	} else {
-		t.Logf("hostname %q does not contain short id %q (non-fatal)", hn, shortID)
-	}
 }
 
 // runConsoleOutput verifies GetConsoleOutput round-trips the
