+++ tests/e2e/single/lifecycle_test.go
@@ -7,7 +7,6 @@ import (
 	"os"
 	"os/exec"
 	"strconv"
-	"strings"
 	"sync/atomic"
 	"testing"
 	"time"
@@ -15,7 +14,6 @@ import (
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/ec2"
 	"github.com/mulgadc/spinifex/tests/e2e/harness"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
 
@@ -160,204 +158,6 @@ func runAttachToStoppedError(t *testing.T, fix *Fixture) {
 	cleaned = true
 }
 
-// runModifyInstanceAttribute changes the (stopped) instance type to
-// the same-family ".small" upsize, starts the instance, SSHes in to
-// verify the new vCPU + memory budget, then stops the instance again so
-// Phase 7c-pre can drive its own start/reboot cycle. Maps to run-e2e.sh
-// ~1070–1177.
-func runModifyInstanceAttribute(t *testing.T, fix *Fixture) {
-	harness.Phase(t, "Single — ModifyInstanceAttribute")
-
-	inst, _ := needInstance(t, fix)
-	instanceID := aws.StringValue(inst.InstanceId)
-	origType, _ := needInstanceTypeArch(t, fix)
-	_, keyPath := needKeyPair(t, fix)
-
-	// Bash strips the ".nano" suffix and appends ".small" — same family,
-	// more RAM at matching vCPU. Avoids xlarge (16 GiB) which the CI host
-	// can't satisfy.
-	if !strings.HasSuffix(origType, ".nano") {
-		t.Fatalf("phase7b: expected discovered instance type to end with .nano, got %q", origType)
-	}
-	modifyType := strings.TrimSuffix(origType, ".nano") + ".small"
-	harness.Detail(t, "from_type", origType, "to_type", modifyType)
-
-	// Stop the instance first — ModifyInstanceAttribute on a running
-	// instance is rejected. Restore (type + state) at end so sibling Test*
-	// see the canonical running singleton.
-	harness.Step(t, "stop-instances %s (precondition for modify)", instanceID)
-	_, err := fix.AWS.EC2.StopInstances(&ec2.StopInstancesInput{
-		InstanceIds: []*string{aws.String(instanceID)},
-	})
-	require.NoError(t, err, "stop-instances")
-	harness.WaitForInstanceState(t, fix.AWS, instanceID, "stopped")
-	t.Cleanup(func() {
-		_, _ = fix.AWS.EC2.StopInstances(&ec2.StopInstancesInput{
-			InstanceIds: []*string{aws.String(instanceID)},
-		})
-		harness.WaitForInstanceState(t, fix.AWS, instanceID, "stopped")
-		_, _ = fix.AWS.EC2.ModifyInstanceAttribute(&ec2.ModifyInstanceAttributeInput{
-			InstanceId:   aws.String(instanceID),
-			InstanceType: &ec2.AttributeValue{Value: aws.String(origType)},
-		})
-		_, _ = fix.AWS.EC2.StartInstances(&ec2.StartInstancesInput{
-			InstanceIds: []*string{aws.String(instanceID)},
-		})
-		harness.WaitForInstanceState(t, fix.AWS, instanceID, "running")
-	})
-
-	// Look up expected vCPUs / memory for the upsized type so we can
-	// assert SSH-reported values match what the AWS surface advertises.
-	typesOut, err := fix.AWS.EC2.DescribeInstanceTypes(&ec2.DescribeInstanceTypesInput{})
-	require.NoError(t, err, "describe-instance-types")
-	var expectedVCPUs int64
-	var expectedMemMiB int64
-	for _, it := range typesOut.InstanceTypes {
-		if aws.StringValue(it.InstanceType) != modifyType {
-			continue
-		}
-		if it.VCpuInfo != nil {
-			expectedVCPUs = aws.Int64Value(it.VCpuInfo.DefaultVCpus)
-		}
-		if it.MemoryInfo != nil {
-			expectedMemMiB = aws.Int64Value(it.MemoryInfo.SizeInMiB)
-		}
-		break
-	}
-	require.NotZero(t, expectedVCPUs, "%s missing VCpuInfo.DefaultVCpus", modifyType)
-	require.NotZero(t, expectedMemMiB, "%s missing MemoryInfo.SizeInMiB", modifyType)
-	harness.Detail(t, "expected_vcpus", expectedVCPUs, "expected_mem_mib", expectedMemMiB)
-
-	harness.Step(t, "modify-instance-attribute %s type=%s", instanceID, modifyType)
-	_, err = fix.AWS.EC2.ModifyInstanceAttribute(&ec2.ModifyInstanceAttributeInput{
-		InstanceId:   aws.String(instanceID),
-		InstanceType: &ec2.AttributeValue{Value: aws.String(modifyType)},
-	})
-	require.NoError(t, err, "modify-instance-attribute")
-
-	// Verify describe-instances reflects the new type before we start.
-	descOut, err := fix.AWS.EC2.DescribeInstances(&ec2.DescribeInstancesInput{
-		InstanceIds: []*string{aws.String(instanceID)},
-	})
-	require.NoError(t, err, "describe-instances")
-	require.NotEmpty(t, descOut.Reservations, "no reservations for %s", instanceID)
-	require.NotEmpty(t, descOut.Reservations[0].Instances, "no instances for %s", instanceID)
-	gotType := aws.StringValue(descOut.Reservations[0].Instances[0].InstanceType)
-	require.Equalf(t, modifyType, gotType,
-		"ModifyInstanceAttribute did not stick: want %s got %s", modifyType, gotType)
-
-	harness.Step(t, "start-instances %s", instanceID)
-	_, err = fix.AWS.EC2.StartInstances(&ec2.StartInstancesInput{
-		InstanceIds: []*string{aws.String(instanceID)},
-	})
-	require.NoError(t, err, "start-instances")
-
-	runInst := harness.WaitForInstanceState(t, fix.AWS, instanceID, "running")
-
-	// Re-discover SSH endpoint — qemu hostfwd may have rebound.
-	host, port := harness.InstancePublicSSHHost(t, runInst)
-	harness.Detail(t, "ssh_host", host, "ssh_port", port)
-
-	waitForSSHReady(t, host, port, keyPath)
-	tgt := harness.SSHTarget{User: "ubuntu", Host: host, Port: port, KeyPath: keyPath}
-
-	harness.Step(t, "ssh nproc")
-	nprocOut := strings.TrimSpace(runSSH(t, tgt, "nproc"))
-	vmVCPUs, err := strconv.ParseInt(nprocOut, 10, 64)
-	require.NoErrorf(t, err, "parse nproc output %q", nprocOut)
-	require.Equalf(t, expectedVCPUs, vmVCPUs,
-		"vCPU mismatch after modify: VM=%d expected=%d", vmVCPUs, expectedVCPUs)
-	harness.Detail(t, "vm_vcpus", vmVCPUs)
-
-	harness.Step(t, "ssh MemTotal")
-	memOut := strings.TrimSpace(runSSH(t, tgt, "awk '/MemTotal/ {print $2}' /proc/meminfo"))
-	vmMemKB, err := strconv.ParseInt(memOut, 10, 64)
-	require.NoErrorf(t, err, "parse MemTotal output %q", memOut)
-	vmMemMiB := vmMemKB / 1024
-	// 15% margin for kernel-reserved memory, matching the bash threshold.
-	expectedMemLow := expectedMemMiB * 85 / 100
-	require.GreaterOrEqualf(t, vmMemMiB, expectedMemLow,
-		"memory too low after modify: VM=%d MiB expected>=%d MiB (target %d MiB)",
-		vmMemMiB, expectedMemLow, expectedMemMiB)
-	harness.Detail(t, "vm_mem_mib", vmMemMiB, "threshold_mib", expectedMemLow)
-
-	// Cleanup restores original type + running state for sibling Test*.
-}
-
-// runRebootInstance starts the (stopped) primary instance, captures its
-// pre-reboot private IP, issues a reboot, asserts the API never reports a
-// non-running state during a short polling window, waits for SSH to come
-// back, and verifies the guest actually rebooted (uptime < 120s) without
-// changing its private IP. Leaves the instance running. Maps to
-// run-e2e.sh ~1178–1236.
-func runRebootInstance(t *testing.T, fix *Fixture) {
-	harness.Phase(t, "Single — Reboot Running Instance")
-
-	inst, _ := needInstance(t, fix)
-	instanceID := aws.StringValue(inst.InstanceId)
-	_, keyPath := needKeyPair(t, fix)
-
-	host, port := harness.InstancePublicSSHHost(t, inst)
-	waitForSSHReady(t, host, port, keyPath)
-
-	// Capture pre-reboot private IP for the post-reboot identity check.
-	preDesc, err := fix.AWS.EC2.DescribeInstances(&ec2.DescribeInstancesInput{
-		InstanceIds: []*string{aws.String(instanceID)},
-	})
-	require.NoError(t, err, "describe-instances (pre-reboot)")
-	require.NotEmpty(t, preDesc.Reservations, "no reservations for %s", instanceID)
-	require.NotEmpty(t, preDesc.Reservations[0].Instances, "no instances for %s", instanceID)
-	preRebootIP := aws.StringValue(preDesc.Reservations[0].Instances[0].PrivateIpAddress)
-	harness.Detail(t, "pre_reboot_private_ip", preRebootIP)
-
-	harness.Step(t, "reboot-instances %s", instanceID)
-	_, err = fix.AWS.EC2.RebootInstances(&ec2.RebootInstancesInput{
-		InstanceIds: []*string{aws.String(instanceID)},
-	})
-	require.NoError(t, err, "reboot-instances")
-
-	// Bash polls 10× at 1s checking the state stays "running" — EC2's
-	// reboot semantics don't transition the instance state at all.
-	for i := 0; i < 10; i++ {
-		out, err := fix.AWS.EC2.DescribeInstances(&ec2.DescribeInstancesInput{
-			InstanceIds: []*string{aws.String(instanceID)},
-		})
-		require.NoError(t, err, "describe-instances during reboot poll %d", i)
-		require.NotEmpty(t, out.Reservations[0].Instances, "instance disappeared during reboot")
-		state := aws.StringValue(out.Reservations[0].Instances[0].State.Name)
-		require.Equalf(t, "running", state,
-			"instance unexpectedly left running state during reboot: %s", state)
-		time.Sleep(1 * time.Second)
-	}
-
-	// SSH endpoint may have rebound after the guest restart (qemu hostfwd
-	// can shift), so re-discover.
-	descPost, err := fix.AWS.EC2.DescribeInstances(&ec2.DescribeInstancesInput{
-		InstanceIds: []*string{aws.String(instanceID)},
-	})
-	require.NoError(t, err, "describe-instances (post-reboot)")
-	require.NotEmpty(t, descPost.Reservations[0].Instances, "no instances post-reboot")
-	postInst := descPost.Reservations[0].Instances[0]
-	host, port = harness.InstancePublicSSHHost(t, postInst)
-	waitForSSHReady(t, host, port, keyPath)
-
-	tgt := harness.SSHTarget{User: "ubuntu", Host: host, Port: port, KeyPath: keyPath}
-	harness.Step(t, "ssh uptime")
-	uptimeOut := strings.TrimSpace(runSSH(t, tgt, "cat /proc/uptime | cut -d. -f1"))
-	uptimeSecs, err := strconv.ParseInt(uptimeOut, 10, 64)
-	require.NoErrorf(t, err, "parse uptime output %q", uptimeOut)
-	require.LessOrEqualf(t, uptimeSecs, int64(120),
-		"guest uptime %ds is > 120s — reboot may not have occurred", uptimeSecs)
-	harness.Detail(t, "uptime_secs", uptimeSecs)
-
-	postRebootIP := aws.StringValue(postInst.PrivateIpAddress)
-	assert.Equalf(t, preRebootIP, postRebootIP,
-		"PrivateIpAddress changed across reboot: %s -> %s", preRebootIP, postRebootIP)
-
-	// Leave instance running — later phases expect the primary instance to
-	// still be up.
-}
-
 // waitForSSHReady probes a full SSH handshake (BatchMode + ConnectTimeout)
 // against host:port, retrying until the daemon completes banner exchange.
 // TCP-port reachability alone is insufficient — sshd accepts the connect
@@ -376,7 +176,7 @@ func waitForSSHReady(t *testing.T, host string, port int, keyPath string) {
 	}
 }
 
-// waitForSSHHandshake is an alias kept for instance_test.go's call site.
+// waitForSSHHandshake is an alias kept for guestchurn_test.go's call site.
 func waitForSSHHandshake(t *testing.T, host string, port int, keyPath string) {
 	waitForSSHReady(t, host, port, keyPath)
 }
